Refactor SiwftObjectStore test so it can be re-used more easily
